Corrected column type of score to float

+++ db/migrate/20190815152914_add_score_to_employee.rb
@@ -1,5 +1,5 @@
 class AddScoreToEmployee < ActiveRecord::Migration[5.2]
   def change
-    add_column :employees, :score, :integer
+    add_column :employees, :score, :float
   end
 end
